+++ lib/Braintree/Digest.php
@@ -12,6 +12,7 @@ class Digest
 {
     public static function hexDigestSha1($key, $string)
     {
+        self::_assertKeyNotNull($key);
         if (function_exists('hash_hmac')) {
             return self::_builtInHmacSha1($string, $key);
         } else {
@@ -58,6 +59,13 @@ class Digest
         }
 
         return sha1($outerPad . pack($pack, sha1($innerPad . $message)));
+    }
+
+    private static function _assertKeyNotNull($key)
+    {
+        if (is_null($key)) {
+            throw new Exception\Configuration("HMAC key must not be null - privateKey is not configured.");
+        }
     }
 }
 // phpcs:enable